chore: move actual_picture to pic module
