+++ src/main.rs
@@ -6,7 +6,6 @@
 extern crate ggez;
 extern crate image;
 
-mod actual_picture;
 mod apt;
 mod parser;
 mod pic;
@@ -23,8 +22,7 @@ use std::sync::{Arc, RwLock};
 use std::thread::spawn;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
-use crate::actual_picture::ActualPicture;
-use crate::pic::{lisp_to_pic, CoordinateSystem, Pic, DEFAULT_COORDINATE_SYSTEM};
+use crate::pic::{lisp_to_pic, CoordinateSystem, Pic, DEFAULT_COORDINATE_SYSTEM, actual_picture::ActualPicture};
 use crate::ui::{Button, MouseButtonState, MouseState, imgui_wrapper::{ImGuiWrapper, EXEC_NAME}};
 
 use clap::Parser;
